fix(cli): re-sanitise pre-shaped mcp__ tool names; split the Cursor adapter tests under the file-size cap
